requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings>=2.10.1

# LLM inference (llama-cpp-python installed with CMAKE_ARGS in Dockerfile)
llama-cpp-python==0.2.90
huggingface-hub==0.35.0

# RAG stack
langchain==0.3.27
langchain-community==0.3.29
langchain-core>=0.3.72,<1.0.0

# Embeddings
sentence-transformers==5.1.0

# Vector store
chromadb==1.1.0

# PDF parsing
pymupdf==1.26.4

# Tokenization
tiktoken==0.11.0

# Utilities
tabulate==0.9.0

# Required for sentence-transformers TF/Keras compatibility
tf-keras
